refactor(core): modularize widget renderers and optimize editor rendering
- Extract individual widget renderers into features/widgets/renderers using Strategy pattern
- Reduce WidgetRenderer.ts from 2070 to ~220 lines with WidgetRegistry
- Create ProfileSvgService for unified SVG delivery, ETag caching and 304 handling
- Add profile metrics telemetry foundation with async non-blocking after() hooks
- Componentize WidgetLibrary into subcomponents (Card, Filters, Templates, Kits)
- Extract WidgetNode, CanvasAlignmentGuides and CanvasMarquee from SVGCanvas
- Optimize editor store subscriptions with fine-grained atomic selectors
