+++ src/engine/core/WidgetRenderer.ts
@@ -1,128 +1,5 @@
-import { APP_DOMAIN, EXTERNAL_LINKS, SOCIAL_PLATFORMS, WIDGET_IDS } from '../../constants'
-import { renderAsciiHeatmap } from '../../features/widgets/renderers/AsciiHeatmapRenderer'
-import { renderAsciiInfoCard } from '../../features/widgets/renderers/AsciiInfoCardRenderer'
-import { renderAsciiPortrait } from '../../features/widgets/renderers/AsciiPortraitRenderer'
-import { renderBentoGrid } from '../../features/widgets/renderers/BentoGridRenderer'
-import { renderBlueprint } from '../../features/widgets/renderers/BlueprintRenderer'
-import { renderCartograph } from '../../features/widgets/renderers/CartographRenderer'
-import { renderCipherPrint } from '../../features/widgets/renderers/CipherPrintRenderer'
-import { renderCodewebHeroOrbit } from '../../features/widgets/renderers/CodewebHeroOrbitRenderer'
-import { renderCodewebMinimalBadge } from '../../features/widgets/renderers/CodewebMinimalBadgeRenderer'
-import { renderCodewebRetroGrid } from '../../features/widgets/renderers/CodewebRetroGridRenderer'
-import { renderCodewebShowcaseCards } from '../../features/widgets/renderers/CodewebShowcaseCardsRenderer'
-import { renderCodewebSocialBadge } from '../../features/widgets/renderers/CodewebSocialBadgeRenderer'
-import { renderCommandDeck } from '../../features/widgets/renderers/CommandDeckRenderer'
-import { renderConstellation } from '../../features/widgets/renderers/ConstellationRenderer'
-import { renderEditorial } from '../../features/widgets/renderers/EditorialRenderer'
-import { renderFieldSpecimen } from '../../features/widgets/renderers/FieldSpecimenRenderer'
-import { renderFoundry } from '../../features/widgets/renderers/FoundryRenderer'
-import { renderGlobe } from '../../features/widgets/renderers/GlobeRenderer'
-import { renderInterlace } from '../../features/widgets/renderers/InterlaceRenderer'
-import { renderMarquee } from '../../features/widgets/renderers/MarqueeRenderer'
-import { renderMetroMap } from '../../features/widgets/renderers/MetroMapRenderer'
-import { renderMonolith } from '../../features/widgets/renderers/MonolithRenderer'
-import { renderNeural } from '../../features/widgets/renderers/NeuralRenderer'
-import { renderPatchbay } from '../../features/widgets/renderers/PatchbayRenderer'
-import { renderPokemonCard } from '../../features/widgets/renderers/PokemonCardRenderer'
-import { renderSignalGrid } from '../../features/widgets/renderers/SignalGridRenderer'
-import { renderSystemLoop } from '../../features/widgets/renderers/SystemLoopRenderer'
-import { renderTerminal } from '../../features/widgets/renderers/TerminalRenderer'
-import { renderTrophies } from '../../features/widgets/renderers/TrophiesRenderer'
-import { renderWakaTime } from '../../features/widgets/renderers/WakaTimeRenderer'
-import { API_ENDPOINTS } from '../../services/endpoints'
-import { generateAsciiArt } from '../ascii/converter'
-import { type AsciiFontName, convertTextToAscii } from '../ascii/textConverter'
 import type { GlobalStyles, NormalizedGitHubData, WidgetInstance } from '../types'
-
-function escapeXml(str: string): string {
-  return str
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&apos;')
-}
-
-function formatUptime(createdAtStr?: string): string {
-  if (!createdAtStr) return '5 years, 3 months, 13 days'
-  const created = new Date(createdAtStr)
-  if (isNaN(created.getTime())) return '5 years, 3 months, 13 days'
-  const now = new Date()
-
-  let years = now.getFullYear() - created.getFullYear()
-  let months = now.getMonth() - created.getMonth()
-  let days = now.getDate() - created.getDate()
-
-  if (days < 0) {
-    months -= 1
-    const prevMonth = new Date(now.getFullYear(), now.getMonth(), 0)
-    days += prevMonth.getDate()
-  }
-
-  if (months < 0) {
-    years -= 1
-    months += 12
-  }
-
-  const parts = []
-  if (years > 0) parts.push(`${years} ${years === 1 ? 'year' : 'years'}`)
-  if (months > 0) parts.push(`${months} ${months === 1 ? 'month' : 'months'}`)
-  if (days > 0 || parts.length === 0) parts.push(`${days} ${days === 1 ? 'day' : 'days'}`)
-
-  return parts.join(', ')
-}
-
-function renderExternalWidgetSvg(
-  url: string,
-  width: number,
-  height: number,
-  title: string,
-  showTitle: boolean,
-  globalStyles: GlobalStyles,
-  accent: string,
-  mode: 'contain' | 'badge' = 'contain',
-  targetUrl?: string,
-  fallbackUrl?: string
-): string {
-  let processedUrl = url
-  try {
-    const parsed = new URL(processedUrl)
-    if (parsed.hostname.toLowerCase() === 'github.com' && parsed.pathname.includes('/blob/')) {
-      processedUrl = processedUrl.replace(
-        /^https?:\/\/github\.com\/([^\/]+)\/([^\/]+)\/blob\/(.+)$/i,
-        'https://raw.githubusercontent.com/$1/$2/$3'
-      )
-    }
-  } catch {}
-
-  const imgY = showTitle ? 44 : 16
-  const paddingX = 16
-  const imgW = width - paddingX * 2
-  const imgH = Math.max(28, height - imgY - 16)
-
-  const imgStyle =
-    mode === 'badge'
-      ? 'height:32px; width:auto; max-width:100%; object-fit:contain; object-position:left center;'
-      : 'width:100%; height:100%; max-width:100%; max-height:100%; object-fit:contain; object-position:left top;'
-
-  const imgHtml = fallbackUrl
-    ? `<img src="${escapeXml(processedUrl)}" alt="${escapeXml(title)}" style="${imgStyle}" onerror="this.onerror=null;this.src='${escapeXml(fallbackUrl)}';" />`
-    : `<img src="${escapeXml(processedUrl)}" alt="${escapeXml(title)}" style="${imgStyle}" />`
-  const innerContentHtml = targetUrl
-    ? `<a href="${escapeXml(targetUrl)}" target="_blank" rel="noopener noreferrer" style="display:inline-block;max-width:100%;max-height:100%;">${imgHtml}</a>`
-    : imgHtml
-
-  return `
-    ${showTitle ? `<text x="24" y="32" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">${escapeXml(title)}</text>` : ''}
-    <!-- EXTERNAL_WIDGET_START: ${escapeXml(processedUrl)} | ${paddingX} | ${imgY} | ${imgW} | ${imgH} | ${mode} | ${fallbackUrl ? escapeXml(fallbackUrl) : ''} -->
-    <foreignObject x="${paddingX}" y="${imgY}" width="${imgW}" height="${imgH}">
-      <div xmlns="http://www.w3.org/1999/xhtml" style="width:100%;height:100%;display:flex;align-items:flex-start;justify-content:flex-start;overflow:hidden;">
-        ${innerContentHtml}
-      </div>
-    </foreignObject>
-    <!-- EXTERNAL_WIDGET_END -->
-  `
-}
+import { getRenderer } from './WidgetRegistry'
 
 export function getWidgetMinSize(
   widget: WidgetInstance,
@@ -179,1606 +56,8 @@ export function renderWidgetSvg(
   const accent = (cfg.accentColor as string) || globalStyles.accentColor || '#c5ff4a'
   const rx = cfg.borderRadius !== undefined ? cfg.borderRadius : globalStyles.borderRadius || 0
 
-  let contentSvg = ''
-
-  switch (widget.widgetId) {
-    case 'header': {
-      const name = escapeXml(data.user.name || data.user.login)
-      const handle = escapeXml(`@${data.user.login}`)
-      const company = data.user.company ? escapeXml(data.user.company) : ''
-
-      contentSvg = `
-        <text x="24" y="44" font-family="${globalStyles.fontFamily}" font-size="28" font-weight="300" fill="${textClr}">${name}</text>
-        <text x="24" y="72" font-family="'JetBrains Mono', monospace" font-size="14" fill="${accent}">${handle}</text>
-        ${company ? `<text x="${width - 24}" y="44" text-anchor="end" font-family="'Inter Tight', sans-serif" font-size="12" fill="#7a7a7a">[ ${company} ]</text>` : ''}
-      `
-      break
-    }
-
-    case 'avatar': {
-      const sourceType = (cfg.sourceType as 'avatar' | 'url' | 'upload') || 'avatar'
-      let avatarUrl = data.user.avatar_url
-
-      if (sourceType === 'upload' && cfg.uploadedImageData) {
-        avatarUrl = cfg.uploadedImageData as string
-      } else if (sourceType === 'url' && cfg.imageUrl) {
-        avatarUrl = cfg.imageUrl as string
-      } else if (cfg.avatarUrl && !cfg.sourceType) {
-        avatarUrl = cfg.avatarUrl as string
-      }
-
-      contentSvg = `
-        <clipPath id="avatar-clip-${widget.instanceId}">
-          <rect class="no-anim" x="16" y="16" width="${width - 32}" height="${height - 32}" rx="${Math.max(4, rx)}" />
-        </clipPath>
-        <rect x="16" y="16" width="${width - 32}" height="${height - 32}" rx="${Math.max(4, rx)}" fill="#060606" stroke="${accent}" stroke-width="1.5" />
-        <image href="${escapeXml(avatarUrl)}" x="16" y="16" width="${width - 32}" height="${height - 32}" preserveAspectRatio="xMidYMid meet" clip-path="url(#avatar-clip-${widget.instanceId})" />
-      `
-      break
-    }
-
-    case 'ascii-art': {
-      const fontSize = Number(cfg.fontSize) || 9
-      const charWidth = fontSize * 0.58
-      const lineHeight = Math.max(7, Math.round(fontSize * 1.12))
-      const colorMode = (cfg.colorMode as string) || 'monochrome'
-      const asciiText = Array.isArray(cfg.asciiText) ? (cfg.asciiText as string[]) : undefined
-      const asciiColors = Array.isArray(cfg.asciiColors)
-        ? (cfg.asciiColors as string[][])
-        : undefined
-
-      const asciiLines =
-        asciiText ||
-        generateAsciiArt(data.user.login, {
-          charset: (cfg.charset as string) || 'dense',
-          customCharset: cfg.customCharset as string,
-          invert: Boolean(cfg.invert),
-          cols: Math.floor((width - 32) / charWidth),
-          rows: Math.floor((height - 32) / lineHeight),
-        })
-
-      const maxCols = Math.max(...asciiLines.map((l) => l.length), 12)
-      const viewW = maxCols * charWidth
-      const viewH = asciiLines.length * lineHeight
-
-      const xCoords = Array.from({ length: maxCols }, (_, i) => (i * charWidth).toFixed(2)).join(
-        ' '
-      )
-      const fontFamily =
-        "'JetBrains Mono', 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier, monospace"
-
-      let innerContent = ''
-
-      if (colorMode === 'color' && asciiColors && asciiColors.length === asciiLines.length) {
-        innerContent = asciiLines
-          .map((line, rowIndex) => {
-            const rowColors = asciiColors[rowIndex] || []
-            let rowSvg = ''
-
-            for (let charIndex = 0; charIndex < line.length;) {
-              let chunk = line[charIndex]
-              const charColor = rowColors[charIndex] || accent
-              let nextIndex = charIndex + 1
-              while (nextIndex < line.length && (rowColors[nextIndex] || accent) === charColor) {
-                chunk += line[nextIndex]
-                nextIndex++
-              }
-              rowSvg += `<tspan fill="${charColor}">${escapeXml(chunk)}</tspan>`
-              charIndex = nextIndex
-            }
-
-            const yPos = (rowIndex + 0.85) * lineHeight
-            return `<text x="${xCoords}" y="${yPos}" font-family="${fontFamily}" font-size="${fontSize}" xml:space="preserve">${rowSvg}</text>`
-          })
-          .join('\n')
-      } else {
-        innerContent = asciiLines
-          .map((line, rowIndex) => {
-            const yPos = (rowIndex + 0.85) * lineHeight
-            return `<text x="${xCoords}" y="${yPos}" font-family="${fontFamily}" font-size="${fontSize}" fill="${accent}" xml:space="preserve">${escapeXml(line)}</text>`
-          })
-          .join('\n')
-      }
-
-      contentSvg = `
-        <svg x="16" y="16" width="${width - 32}" height="${height - 32}" viewBox="0 0 ${viewW} ${viewH}" preserveAspectRatio="xMidYMid meet">
-          ${innerContent}
-        </svg>
-      `
-      break
-    }
-
-    case 'ascii-text': {
-      const fontSize = Number(cfg.fontSize) || 12
-      const charWidth = fontSize * 0.6
-      const lineHeight = fontSize * 1.2
-      const asciiLines = Array.isArray(cfg.asciiLines)
-        ? (cfg.asciiLines as string[])
-        : convertTextToAscii(
-            (cfg.customText as string) || 'GitAscii',
-            (cfg.asciiFont as AsciiFontName) || 'block',
-            cfg.charSpacing !== undefined ? Number(cfg.charSpacing) : 1,
-            (cfg.charset as string) || 'default',
-            (cfg.customCharset as string) || ''
-          )
-
-      const maxCols = Math.max(...asciiLines.map((l) => l.length), 1)
-      const viewW = maxCols * charWidth
-      const viewH = asciiLines.length * lineHeight
-
-      const xCoords = Array.from({ length: maxCols }, (_, i) => (i * charWidth).toFixed(2)).join(
-        ' '
-      )
-      const fontFamily =
-        "'JetBrains Mono', 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier, monospace"
-
-      const linesContent = asciiLines
-        .map((line, rowIndex) => {
-          const yPos = (rowIndex + 0.85) * lineHeight
-          return `<text x="${xCoords}" y="${yPos}" font-family="${fontFamily}" font-size="${fontSize}" fill="${accent}" xml:space="preserve">${escapeXml(line)}</text>`
-        })
-        .join('\n')
-
-      contentSvg = `
-        <svg x="16" y="16" width="${width - 32}" height="${height - 32}" viewBox="0 0 ${viewW} ${viewH}" preserveAspectRatio="xMidYMid meet">
-          ${linesContent}
-        </svg>
-      `
-      break
-    }
-
-    case 'bio': {
-      const customBio =
-        cfg.customBio !== undefined
-          ? (cfg.customBio as string)
-          : data.user.bio || 'No bio provided.'
-      const customLocation =
-        cfg.customLocation !== undefined ? (cfg.customLocation as string) : data.user.location || ''
-      const customBlog =
-        cfg.customBlog !== undefined ? (cfg.customBlog as string) : data.user.blog || ''
-
-      const maxCharsPerLine = Math.max(20, Math.floor((width - 72) / 8.5))
-      const wrappedLines: string[] = []
-
-      for (const p of customBio.split('\n')) {
-        if (p.length <= maxCharsPerLine) {
-          wrappedLines.push(p)
-          continue
-        }
-        let remaining = p
-        while (remaining.length > 0) {
-          if (remaining.length <= maxCharsPerLine) {
-            wrappedLines.push(remaining)
-            break
-          }
-          let breakPoint = remaining.lastIndexOf(' ', maxCharsPerLine)
-          if (breakPoint === -1) {
-            breakPoint = maxCharsPerLine
-          }
-          wrappedLines.push(remaining.substring(0, breakPoint))
-          remaining = remaining.substring(breakPoint + 1).trimStart()
-        }
-      }
-
-      const bioSvg = wrappedLines
-        .map((line, i) => `<tspan x="24" dy="${i === 0 ? 0 : 20}">${escapeXml(line)}</tspan>`)
-        .join('')
-
-      const requiredHeight = 60 + (Math.max(1, wrappedLines.length) - 1) * 20 + 48
-      const finalHeight = Math.max(height, requiredHeight)
-
-      let blogHref = customBlog
-      if (blogHref && !blogHref.startsWith('http://') && !blogHref.startsWith('https://')) {
-        blogHref = `https://${blogHref}`
-      }
-
-      const locationSvg = customLocation
-        ? `<text x="0" y="0" font-family="${globalStyles.fontFamily}" font-size="12" fill="#7a7a7a">📍 ${escapeXml(customLocation)}</text>`
-        : ''
-
-      const blogSvg = customBlog
-        ? `<a href="${escapeXml(blogHref)}" target="_blank" rel="noopener noreferrer" cursor="pointer">
-             <text x="${customLocation ? 180 : 0}" y="0" font-family="${globalStyles.fontFamily}" font-size="12" fill="${accent}" text-decoration="underline">🌐 ${escapeXml(customBlog)}</text>
-           </a>`
-        : ''
-
-      contentSvg = `
-        <text x="24" y="32" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ BIOGRAPHY ]</text>
-        <text x="24" y="60" font-family="${globalStyles.fontFamily}" font-size="14" fill="${textClr}">
-          ${bioSvg}
-        </text>
-        <g transform="translate(24, ${finalHeight - 24})">
-          ${locationSvg}
-          ${blogSvg}
-        </g>
-      `
-      break
-    }
-
-    case 'stats': {
-      const hideMetrics: string[] = Array.isArray(cfg.hideMetrics)
-        ? (cfg.hideMetrics as string[])
-        : []
-
-      const allMetrics = [
-        { id: 'stars', label: 'STARS', val: data.totalStars.toLocaleString() },
-        { id: 'repos', label: 'REPOS', val: data.user.public_repos.toLocaleString() },
-        { id: 'followers', label: 'FOLLOWERS', val: data.user.followers.toLocaleString() },
-        { id: 'following', label: 'FOLLOWING', val: data.user.following.toLocaleString() },
-        { id: 'forks', label: 'FORKS', val: data.totalForks.toLocaleString() },
-        { id: 'gists', label: 'GISTS', val: data.user.public_gists.toLocaleString() },
-      ]
-
-      const statItems = allMetrics.filter((m) => !hideMetrics.includes(m.id))
-      const statsLayout = (cfg.statsLayout as string) || 'horizontal'
-      const statsStyle = (cfg.statsStyle as string) || 'default'
-      const labelStyle = (cfg.labelStyle as string) || 'label'
-      const valueFontSize = Number(cfg.valueFontSize) || 28
-
-      const getLabel = (m: { label: string }) => (labelStyle === 'none' ? '' : m.label)
-
-      let statsSvg = ''
-
-      if (statsStyle === 'terminal') {
-        const startY = 48
-        const monoFont = `'JetBrains Mono', monospace`
-        if (statsLayout === 'horizontal') {
-          const itemW = statItems.length > 0 ? (width - 48) / statItems.length : width - 48
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(${24 + i * itemW}, ${startY})">
-              <text x="0" y="18" font-family="${monoFont}" font-size="13" fill="${accent}">[ ${m.val} ]</text>
-              ${labelStyle !== 'none' ? `<text x="0" y="32" font-family="${monoFont}" font-size="9" fill="#7a7a7a" letter-spacing="2">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        } else if (statsLayout === 'vertical') {
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(24, ${startY + i * 32})">
-              <text x="0" y="18" font-family="${monoFont}" font-size="13" fill="${accent}">[ ${m.val} ]</text>
-              ${labelStyle !== 'none' ? `<text x="${(m.val.length + 4) * 8 + 4}" y="18" font-family="${monoFont}" font-size="9" fill="#7a7a7a" letter-spacing="2">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        } else {
-          const colW = (width - 48) / 2
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(${24 + (i % 2) * colW}, ${startY + Math.floor(i / 2) * 40})">
-              <text x="0" y="18" font-family="${monoFont}" font-size="13" fill="${accent}">[ ${m.val} ]</text>
-              ${labelStyle !== 'none' ? `<text x="0" y="30" font-family="${monoFont}" font-size="9" fill="#7a7a7a" letter-spacing="2">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        }
-      } else if (statsStyle === 'minimal') {
-        const fs = valueFontSize
-        if (statsLayout === 'horizontal') {
-          const itemW = statItems.length > 0 ? (width - 48) / statItems.length : width - 48
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <text x="${24 + i * itemW}" y="${48 + fs}" font-family="${globalStyles.fontFamily}" font-size="${fs}" font-weight="200" fill="${textClr}">${m.val}</text>`
-            )
-            .join('')
-        } else if (statsLayout === 'vertical') {
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <text x="24" y="${48 + i * (fs + 12) + fs}" font-family="${globalStyles.fontFamily}" font-size="${Math.min(fs, 22)}" font-weight="200" fill="${textClr}">${m.val}</text>`
-            )
-            .join('')
-        } else {
-          const colW = (width - 48) / 2
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <text x="${24 + (i % 2) * colW}" y="${48 + Math.floor(i / 2) * (Math.min(fs, 24) + 8) + Math.min(fs, 24)}" font-family="${globalStyles.fontFamily}" font-size="${Math.min(fs, 24)}" font-weight="200" fill="${textClr}">${m.val}</text>`
-            )
-            .join('')
-        }
-      } else if (statsStyle === 'cards') {
-        const cardH = 52
-        const gap = 8
-        const cols = statsLayout === 'vertical' ? 1 : 2
-        const cardW = cols === 1 ? width - 48 : Math.floor((width - 48 - gap) / 2)
-        statsSvg = statItems
-          .map((m, i) => {
-            const col = i % cols
-            const row = Math.floor(i / cols)
-            const cx = 24 + col * (cardW + gap)
-            const cy = 44 + row * (cardH + gap)
-            return `
-            <g transform="translate(${cx}, ${cy})">
-              <rect x="0" y="0" width="${cardW}" height="${cardH}" fill="#1e1e1e" rx="6" />
-              <rect x="0" y="0" width="3" height="${cardH}" fill="${accent}" rx="1" />
-              <text x="12" y="22" font-family="${globalStyles.fontFamily}" font-size="${Math.min(valueFontSize, 22)}" font-weight="300" fill="${accent}">${m.val}</text>
-              ${labelStyle !== 'none' ? `<text x="12" y="42" font-family="${globalStyles.fontFamily}" font-size="9" font-weight="500" fill="#7a7a7a" letter-spacing="1.5">${getLabel(m)}</text>` : ''}
-            </g>`
-          })
-          .join('')
-      } else {
-        if (statsLayout === 'horizontal') {
-          const itemWidth = statItems.length > 0 ? (width - 48) / statItems.length : width - 48
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(${24 + i * itemWidth}, 48)">
-              <text x="0" y="${valueFontSize}" font-family="${globalStyles.fontFamily}" font-size="${valueFontSize}" font-weight="300" fill="${accent}">${m.val}</text>
-              ${labelStyle !== 'none' ? `<text x="0" y="${valueFontSize + 18}" font-family="${globalStyles.fontFamily}" font-size="10" font-weight="500" fill="#7a7a7a" letter-spacing="1.5">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        } else if (statsLayout === 'vertical') {
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(24, ${48 + i * 52})">
-              <text x="0" y="28" font-family="${globalStyles.fontFamily}" font-size="${Math.min(valueFontSize, 24)}" font-weight="300" fill="${accent}">${m.val}</text>
-              ${labelStyle !== 'none' ? `<text x="${Math.min(valueFontSize, 24) * (m.val.length * 0.6) + 8}" y="28" font-family="${globalStyles.fontFamily}" font-size="10" font-weight="500" fill="#7a7a7a" letter-spacing="1.5">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        } else {
-          const colW = (width - 48) / 2
-          statsSvg = statItems
-            .map(
-              (m, i) => `
-            <g transform="translate(${24 + (i % 2) * colW}, ${48 + Math.floor(i / 2) * 60})">
-              <text x="0" y="${Math.min(valueFontSize, 26)}" font-family="${globalStyles.fontFamily}" font-size="${Math.min(valueFontSize, 26)}" font-weight="300" fill="${accent}">${m.val}</text>
-              ${labelStyle !== 'none' ? `<text x="0" y="${Math.min(valueFontSize, 26) + 16}" font-family="${globalStyles.fontFamily}" font-size="10" font-weight="500" fill="#7a7a7a" letter-spacing="1.5">${getLabel(m)}</text>` : ''}
-            </g>`
-            )
-            .join('')
-        }
-      }
-
-      contentSvg = `
-        <text x="24" y="32" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ GITHUB METRICS ]</text>
-        ${statsSvg}
-      `
-      break
-    }
-
-    case 'pokemon-card': {
-      contentSvg = renderPokemonCard(widget, data, globalStyles, width, height)
-      break
-    }
-
-    case 'languages': {
-      const hideLangsArr: string[] = Array.isArray(cfg.hideLangsArr)
-        ? (cfg.hideLangsArr as string[])
-        : []
-      const hideLangsStr =
-        typeof cfg.hideLangs === 'string'
-          ? (cfg.hideLangs as string)
-              .split(',')
-              .map((l) => l.trim().toLowerCase())
-              .filter(Boolean)
-          : []
-      const hideLangs = [...hideLangsArr.map((l) => l.toLowerCase()), ...hideLangsStr]
-
-      let filteredLangs = Object.entries(data.languages)
-      if (hideLangs.length > 0) {
-        filteredLangs = filteredLangs.filter(([lang]) => !hideLangs.includes(lang.toLowerCase()))
-      }
-
-      const maxLangs = Number(cfg.langsCount) || 5
-      const topLangs = filteredLangs.slice(0, maxLangs)
-      const totalCount = topLangs.reduce((sum, [_, count]) => sum + count, 0) || 1
-      const showPercentage = cfg.showPercentage !== false
-      const langsLayout = (cfg.langsLayout as string) || 'bars'
-
-      const langColors: Record<string, string> = {
-        TypeScript: '#3178c6',
-        JavaScript: '#f1e05a',
-        Rust: '#dea584',
-        Python: '#3572A5',
-        CSS: '#563d7c',
-        HTML: '#e34c26',
-        Go: '#00ADD8',
-        Java: '#b07219',
-        Ruby: '#701516',
-        'C++': '#f34b7d',
-        'C#': '#239120',
-        PHP: '#4F5D95',
-        Swift: '#F05138',
-        Kotlin: '#A97BFF',
-        Dart: '#00B4AB',
-        Shell: '#89e051',
-        Vue: '#41b883',
-        Svelte: '#ff3e00',
-      }
-
-      const getColor = (lang: string) => langColors[lang] || accent
-
-      let langsSvg = ''
-
-      if (langsLayout === 'bars' || langsLayout === undefined) {
-        let currentX = 24
-        const barWidth = width - 48
-        const barSvg = topLangs
-          .map(([lang, count]) => {
-            const w = (count / totalCount) * barWidth
-            const rect = `<rect x="${currentX}" y="52" width="${w}" height="8" fill="${getColor(lang)}" rx="2" />`
-            currentX += w
-            return rect
-          })
-          .join('')
-
-        const legendSvg = topLangs
-          .map(([lang, count], i) => {
-            const pct = Math.round((count / totalCount) * 100)
-            return `
-          <g transform="translate(${24 + (i % 2) * (barWidth / 2)}, ${80 + Math.floor(i / 2) * 24})">
-            <circle cx="6" cy="-4" r="4" fill="${getColor(lang)}" />
-            <text x="16" y="0" font-family="'Inter Tight', sans-serif" font-size="12" fill="${textClr}">${lang} ${showPercentage ? `<tspan fill="#7a7a7a">${pct}%</tspan>` : ''}</text>
-          </g>
-        `
-          })
-          .join('')
-
-        langsSvg = `${barSvg}${legendSvg}`
-      } else if (langsLayout === 'list') {
-        const barW = width - 48
-        langsSvg = topLangs
-          .map(([lang, count], i) => {
-            const pct = Math.round((count / totalCount) * 100)
-            const fillW = (count / totalCount) * (barW - 100)
-            return `
-          <g transform="translate(24, ${48 + i * 26})">
-            <circle cx="6" cy="8" r="4" fill="${getColor(lang)}" />
-            <text x="18" y="14" font-family="'Inter Tight', sans-serif" font-size="12" fill="${textClr}">${escapeXml(lang)}</text>
-            ${showPercentage ? `<text x="${barW - 36}" y="14" text-anchor="end" font-family="'Inter Tight', sans-serif" font-size="11" fill="#7a7a7a">${pct}%</text>` : ''}
-            <rect x="0" y="20" width="${barW - 36}" height="3" fill="#252525" rx="1" />
-            <rect x="0" y="20" width="${fillW}" height="3" fill="${getColor(lang)}" rx="1" />
-          </g>
-        `
-          })
-          .join('')
-      } else if (langsLayout === 'compact') {
-        const itemW = (width - 48) / Math.min(topLangs.length, 3)
-        langsSvg = topLangs
-          .map(([lang, count], i) => {
-            const pct = Math.round((count / totalCount) * 100)
-            return `
-          <g transform="translate(${24 + (i % 3) * itemW}, ${48 + Math.floor(i / 3) * 52})">
-            <circle cx="6" cy="8" r="5" fill="${getColor(lang)}" />
-            <text x="16" y="14" font-family="'Inter Tight', sans-serif" font-size="11" fill="${textClr}">${escapeXml(lang)}</text>
-            ${showPercentage ? `<text x="16" y="30" font-family="'Inter Tight', sans-serif" font-size="10" fill="#7a7a7a">${pct}%</text>` : ''}
-          </g>
-        `
-          })
-          .join('')
-      } else if (langsLayout === 'donut') {
-        const donutLegendPos = (cfg.donutLegendPos as string) || 'bottom'
-        const donutShowPct = cfg.donutShowPct !== false
-        const donutCenterLabel = Boolean(cfg.donutCenterLabel)
-
-        const donutR =
-          donutLegendPos === 'side' ? Math.min(width / 4, 60) : Math.min((width - 48) / 2, 60)
-        const donutRi = Math.round(donutR * 0.52)
-        const donutCx = donutLegendPos === 'side' ? 24 + donutR + 4 : width / 2
-        const donutCy = 44 + donutR
-
-        let startAngle = -Math.PI / 2
-        const arcsSvg = topLangs
-          .map(([lang, count]) => {
-            const fraction = count / totalCount
-            const angle = fraction * 2 * Math.PI
-            const endAngle = startAngle + angle
-            const x1 = donutCx + donutR * Math.cos(startAngle)
-            const y1 = donutCy + donutR * Math.sin(startAngle)
-            const x2 = donutCx + donutR * Math.cos(endAngle)
-            const y2 = donutCy + donutR * Math.sin(endAngle)
-            const xi1 = donutCx + donutRi * Math.cos(startAngle)
-            const yi1 = donutCy + donutRi * Math.sin(startAngle)
-            const xi2 = donutCx + donutRi * Math.cos(endAngle)
-            const yi2 = donutCy + donutRi * Math.sin(endAngle)
-            const large = angle > Math.PI ? 1 : 0
-            const path = `M ${x1.toFixed(1)} ${y1.toFixed(1)} A ${donutR} ${donutR} 0 ${large} 1 ${x2.toFixed(1)} ${y2.toFixed(1)} L ${xi2.toFixed(1)} ${yi2.toFixed(1)} A ${donutRi} ${donutRi} 0 ${large} 0 ${xi1.toFixed(1)} ${yi1.toFixed(1)} Z`
-            startAngle = endAngle
-            return `<path d="${path}" fill="${getColor(lang)}" />`
-          })
-          .join('')
-
-        const centerSvg =
-          donutCenterLabel && topLangs.length > 0
-            ? `<text x="${donutCx}" y="${donutCy + 4}" text-anchor="middle" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="600" fill="${textClr}">${escapeXml(topLangs[0][0])}</text>`
-            : ''
-
-        let legendSvg = ''
-        if (donutLegendPos === 'bottom') {
-          const legendStartY = donutCy + donutR + 16
-          const lW = width - 48
-          legendSvg = topLangs
-            .map(([lang, count], i) => {
-              const pct = Math.round((count / totalCount) * 100)
-              return `
-              <g transform="translate(${24 + (i % 2) * (lW / 2)}, ${legendStartY + Math.floor(i / 2) * 22})">
-                <circle cx="6" cy="-4" r="4" fill="${getColor(lang)}" />
-                <text x="16" y="0" font-family="'Inter Tight', sans-serif" font-size="11" fill="${textClr}">${escapeXml(lang)}${donutShowPct ? ` <tspan fill="#7a7a7a">${pct}%</tspan>` : ''}</text>
-              </g>`
-            })
-            .join('')
-        } else if (donutLegendPos === 'side') {
-          const legendX = donutCx + donutR + 16
-          legendSvg = topLangs
-            .slice(0, 6)
-            .map(([lang, count], i) => {
-              const pct = Math.round((count / totalCount) * 100)
-              return `
-              <g transform="translate(${legendX}, ${donutCy - donutR + i * 22})">
-                <circle cx="6" cy="-4" r="4" fill="${getColor(lang)}" />
-                <text x="16" y="0" font-family="'Inter Tight', sans-serif" font-size="10" fill="${textClr}">${escapeXml(lang.length > 10 ? lang.slice(0, 9) + '…' : lang)}${donutShowPct ? ` <tspan fill="#7a7a7a">${pct}%</tspan>` : ''}</text>
-              </g>`
-            })
-            .join('')
-        }
-
-        langsSvg = `${arcsSvg}${centerSvg}${legendSvg}`
-      }
-
-      contentSvg = `
-        <text x="24" y="32" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ TOP LANGUAGES ]</text>
-        ${langsSvg}
-      `
-      break
-    }
-
-    case 'repositories': {
-      const selectedRepos: string[] = Array.isArray(cfg.selectedRepos)
-        ? (cfg.selectedRepos as string[])
-        : []
-      const maxRepos = Number(cfg.maxRepos) || 3
-      const repoViewMode = (cfg.repoViewMode as string) || 'list'
-      const repoSortBy = (cfg.repoSortBy as string) || 'stars'
-      const showRepoLanguage = cfg.showRepoLanguage !== false
-      const showRepoForks = Boolean(cfg.showRepoForks)
-
-      let repoList = [...data.repos].filter((r) => !r.fork)
-
-      if (selectedRepos.length > 0) {
-        const ordered = selectedRepos
-          .map((name) => repoList.find((r) => r.name === name))
-          .filter(Boolean) as typeof repoList
-        const rest = repoList.filter((r) => !selectedRepos.includes(r.name))
-        repoList = [...ordered, ...rest]
-      } else {
-        if (repoSortBy === 'updated') {
-          repoList.sort(
-            (a, b) => new Date(b.updated_at).getTime() - new Date(a.updated_at).getTime()
-          )
-        } else if (repoSortBy === 'forks') {
-          repoList.sort((a, b) => b.forks_count - a.forks_count)
-        } else if (repoSortBy === 'name') {
-          repoList.sort((a, b) => a.name.localeCompare(b.name))
-        } else {
-          repoList.sort((a, b) => b.stargazers_count - a.stargazers_count)
-        }
-      }
-
-      const repos = repoList.slice(0, maxRepos)
-
-      if (repoViewMode === 'grid') {
-        const cols = 2
-        const cardW = Math.floor((width - 48 - 12) / cols)
-        const cardH = 80
-        const gapY = 12
-
-        contentSvg = `
-          <text x="24" y="32" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ FEATURED REPOSITORIES ]</text>
-          ${repos
-            .map(
-              (repo, i) => `
-            <g transform="translate(${24 + (i % cols) * (cardW + 12)}, ${50 + Math.floor(i / cols) * (cardH + gapY)})">
-              <rect x="0" y="0" width="${cardW}" height="${cardH}" fill="#1e1e1e" rx="4" />
-              <rect x="0" y="0" width="4" height="${cardH}" fill="${accent}" rx="2" />
-              <text x="12" y="20" font-family="'JetBrains Mono', monospace" font-size="12" font-weight="500" fill="${accent}">${escapeXml(repo.name.length > 18 ? repo.name.slice(0, 16) + '…' : repo.name)}</text>
-              <text x="${cardW - 8}" y="20" text-anchor="end" font-family="'Inter Tight', sans-serif" font-size="11" fill="#7a7a7a">★ ${repo.stargazers_count}</text>
-              <text x="12" y="38" font-family="'Inter Tight', sans-serif" font-size="10" fill="#7a7a7a">${escapeXml((repo.description || 'No description.').slice(0, 40) + ((repo.description || '').length > 40 ? '…' : ''))}</text>
-              ${showRepoLanguage && repo.language ? `<text x="12" y="${cardH - 10}" font-family="'Inter Tight', sans-serif" font-size="10" fill="${accent}">${escapeXml(repo.language)}</text>` : ''}
-              ${showRepoForks ? `<text x="${cardW - 8}" y="${cardH - 10}" text-anchor="end" font-family="'Inter Tight', sans-serif" font-size="10" fill="#7a7a7a">⑂ ${repo.forks_count}</text>` : ''}
-            </g>
-          `
-            )
-            .join('')}
-        `
-      } else {
-        const showRepoStars = cfg.showRepoStars !== false
-        const showRepoDesc = cfg.showRepoDesc !== false
-        const showRepoUpdated = Boolean(cfg.showRepoUpdated)
-
-        const metaLineNeeded = showRepoLanguage || showRepoForks || showRepoStars || showRepoUpdated
-        const cardH = 24 + (showRepoDesc ? 18 : 0) + (metaLineNeeded ? 18 : 0) + 8
-        const rowSpacing = cardH + 8
-
-        contentSvg = `
-          <text x="24" y="32" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ FEATURED REPOSITORIES ]</text>
-          ${repos
-            .map((repo, i) => {
-              const gy = 50 + i * rowSpacing
-              const metaParts: string[] = []
-              if (showRepoLanguage && repo.language) metaParts.push(repo.language)
-              if (showRepoStars) metaParts.push(`\u2605 ${repo.stargazers_count}`)
-              if (showRepoForks) metaParts.push(`\u2442 ${repo.forks_count}`)
-              if (showRepoUpdated) {
-                const d = new Date(repo.updated_at)
-                metaParts.push(
-                  `Updated ${d.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })}`
-                )
-              }
-              const metaStr = escapeXml(metaParts.join('  '))
-
-              let yOff = 20
-              const nameRow = `<text x="12" y="${yOff}" font-family="'JetBrains Mono', monospace" font-size="12" font-weight="500" fill="${accent}">${escapeXml(repo.name.length > 36 ? repo.name.slice(0, 34) + '\u2026' : repo.name)}</text>`
-              yOff += showRepoDesc ? 18 : 0
-              const descRow = showRepoDesc
-                ? `<text x="12" y="${yOff}" font-family="'Inter Tight', sans-serif" font-size="10" fill="#7a7a7a">${escapeXml((repo.description || 'No description.').slice(0, 68) + ((repo.description || '').length > 68 ? '\u2026' : ''))}</text>`
-                : ''
-              yOff += metaLineNeeded ? 18 : 0
-              const metaRow = metaLineNeeded
-                ? `<text x="12" y="${yOff}" font-family="'Inter Tight', sans-serif" font-size="10" fill="${accent}">${metaStr}</text>`
-                : ''
-
-              return `
-            <g transform="translate(24, ${gy})">
-              <rect x="0" y="0" width="${width - 48}" height="${cardH}" fill="#1e1e1e" rx="4" />
-              <rect x="0" y="0" width="3" height="${cardH}" fill="${accent}" rx="1" />
-              ${nameRow}${descRow}${metaRow}
-            </g>`
-            })
-            .join('')}
-        `
-      }
-      break
-    }
-
-    case 'divider': {
-      contentSvg = `<line x1="0" y1="${height / 2}" x2="${width}" y2="${height / 2}" stroke="${accent}" stroke-width="4" />`
-      break
-    }
-
-    case 'footer': {
-      contentSvg = `
-        <rect x="0" y="0" width="${width}" height="${height}" fill="#000000" />
-        <text x="24" y="${height / 2 + 4}" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ GENERATED BY GITASCII ]</text>
-        <text x="${width - 24}" y="${height / 2 + 4}" text-anchor="end" font-family="'JetBrains Mono', monospace" font-size="11" fill="${accent}">${APP_DOMAIN}/${data.user.login}</text>
-      `
-      break
-    }
-
-    case 'tech-stack': {
-      const selectedTechs =
-        Array.isArray(cfg.selectedTechs) && cfg.selectedTechs.length > 0
-          ? (cfg.selectedTechs as string[])
-          : [
-              'js',
-              'ts',
-              'react',
-              'nextjs',
-              'nodejs',
-              'tailwind',
-              'python',
-              'docker',
-              'git',
-              'postgres',
-            ]
-
-      const theme = (cfg.theme as string) || 'dark'
-      const perLine = Number(cfg.perLine) || 12
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ TECHNOLOGIES & SKILLS ]'
-
-      const mappedTechs = selectedTechs.map((t) => (t === 'reactnative' ? 'react' : t))
-      const uniqueTechs = Array.from(new Set(mappedTechs))
-      const techString = uniqueTechs.join(',')
-      const skillIconsUrl = API_ENDPOINTS.SKILL_ICONS.GET(techString, theme, perLine)
-
-      const titleY = 32
-      const imageY = showTitle ? 44 : 16
-      const imageWidth = width - 48
-      const imageHeight = Math.max(40, height - imageY - 16)
-
-      contentSvg = `
-        ${showTitle ? `<text x="24" y="${titleY}" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">${escapeXml(customTitle)}</text>` : ''}
-        <image href="${escapeXml(skillIconsUrl)}" x="24" y="${imageY}" width="${imageWidth}" height="${imageHeight}" preserveAspectRatio="xMinYMin meet" />
-      `
-      break
-    }
-
-    case 'social-media': {
-      const selectedSocials =
-        Array.isArray(cfg.selectedSocials) && cfg.selectedSocials.length > 0
-          ? (cfg.selectedSocials as string[])
-          : ['github', 'linkedin', 'twitter', 'discord', 'youtube', 'website']
-
-      const socialUrls = (cfg.socialUrls as Record<string, string>) || {}
-      const badgeStyle = (cfg.badgeStyle as string) || 'for-the-badge'
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ SOCIAL MEDIA ]'
-      const theme = (cfg.theme as string) || 'dark'
-
-      const titleY = 32
-      const startY = showTitle ? 44 : 16
-
-      if (badgeStyle === 'skillicons') {
-        const socialTechString = selectedSocials.join(',')
-        const skillIconsUrl = API_ENDPOINTS.SKILL_ICONS.GET(socialTechString, theme, 12)
-        const imageWidth = width - 48
-        const imageHeight = Math.max(40, height - startY - 16)
-
-        contentSvg = `
-          ${showTitle ? `<text x="24" y="${titleY}" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">${escapeXml(customTitle)}</text>` : ''}
-          <image href="${escapeXml(skillIconsUrl)}" x="24" y="${startY}" width="${imageWidth}" height="${imageHeight}" preserveAspectRatio="xMinYMin meet" />
-        `
-      } else {
-        const socialPlatformsMap = SOCIAL_PLATFORMS.reduce<
-          Record<string, (typeof SOCIAL_PLATFORMS)[number]>
-        >((acc, curr) => {
-          acc[curr.id] = curr
-          return acc
-        }, {})
-
-        const badgeH = badgeStyle === 'for-the-badge' ? 28 : 22
-        const gapX = 10
-        const gapY = 10
-        const maxX = width - 24
-
-        let currentX = 24
-        let currentY = startY
-
-        const badgesSvg = selectedSocials
-          .map((platformId) => {
-            const p = socialPlatformsMap[platformId]
-            if (!p) return ''
-
-            const label = p.label
-            const badgeW =
-              badgeStyle === 'for-the-badge'
-                ? Math.max(64, Math.round(54 + label.length * 7.6))
-                : Math.max(50, Math.round(40 + label.length * 6.2))
-
-            if (currentX + badgeW > maxX && currentX > 24) {
-              currentX = 24
-              currentY += badgeH + gapY
-            }
-
-            const posX = currentX
-            currentX += badgeW + gapX
-
-            const badgeUrl = API_ENDPOINTS.SHIELDS_IO.CUSTOM_BADGE(
-              label,
-              p.color,
-              badgeStyle,
-              p.logo
-            )
-            const targetUrl =
-              socialUrls[platformId] || p.defaultUrl.replace('{username}', data.user.login)
-
-            return `
-            <a href="${escapeXml(targetUrl)}" target="_blank" rel="noopener noreferrer" cursor="pointer">
-              <image href="${escapeXml(badgeUrl)}" x="${posX}" y="${currentY}" width="${badgeW}" height="${badgeH}" preserveAspectRatio="xMinYMid meet" />
-            </a>
-          `
-          })
-          .filter(Boolean)
-          .join('\n')
-
-        contentSvg = `
-          ${showTitle ? `<text x="24" y="${titleY}" font-family="${globalStyles.fontFamily}" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">${escapeXml(customTitle)}</text>` : ''}
-          ${badgesSvg}
-        `
-      }
-      break
-    }
-
-    case 'terminal-info':
-    case 'terminal-card': {
-      const showMainSection = cfg.showMainSection !== false
-      const showContactSection = cfg.showContactSection !== false
-      const showStatsSection = cfg.showStatsSection !== false
-
-      const showUptime = cfg.showUptime !== false
-      const showLocation = cfg.showLocation !== false
-      const showCompany = cfg.showCompany !== false
-      const showLanguages = cfg.showLanguages !== false
-      const showJoined = Boolean(cfg.showJoined)
-      const showStatus = Boolean(cfg.showStatus)
-      const showPronouns = Boolean(cfg.showPronouns)
-      const showTimezone = Boolean(cfg.showTimezone)
-      const showAchievements = Boolean(cfg.showAchievements)
-      const showHighlights = Boolean(cfg.showHighlights)
-
-      const showWebsite = cfg.showWebsite !== false
-      const showGithub = cfg.showGithub !== false
-      const showTwitter = Boolean(cfg.showTwitter)
-      const showEmail = Boolean(cfg.showEmail)
-      const showOrgs = Boolean(cfg.showOrgs)
-
-      const showRepos = cfg.showRepos !== false
-      const showStars = cfg.showStars !== false
-      const showCommits = cfg.showCommits !== false
-      const showFollowers = cfg.showFollowers !== false
-      const showFollowing = Boolean(cfg.showFollowing)
-      const showGists = Boolean(cfg.showGists)
-
-      const dotLeaders = cfg.dotLeaders !== false
-
-      const headerClr = (cfg.headerColor as string) || '#58a6ff'
-      const labelClr = (cfg.labelColor as string) || '#ffa657'
-      const dotClr = (cfg.dotColor as string) || '#484f58'
-      const valClr = (cfg.valueColor as string) || '#c9d1d9'
-      const statsValClr = (cfg.statsValColor as string) || '#79c0ff'
-      const dividerClr = (cfg.dividerColor as string) || '#3d444d'
-
-      const fontSize = 14
-      const fontCharWidth = fontSize * 0.6
-      const paddingX = 24
-      const totalChars = Math.max(26, Math.floor((width - paddingX * 2) / fontCharWidth))
-      const lineHeight = Math.max(18, Math.floor(fontSize * 1.35))
-
-      const lines: string[] = []
-      let currentY = 28
-
-      if (showMainSection) {
-        const titleStr = (cfg.customTitle as string) || `${data.user.login}@github`
-        const dashesCount = Math.max(2, totalChars - 1 - (titleStr.length + 2))
-        const dashesStr = '─'.repeat(dashesCount)
-
-        lines.push(
-          `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${dividerClr}">─</tspan><tspan fill="${headerClr}"> ${escapeXml(titleStr)} </tspan><tspan fill="${dividerClr}">${dashesStr}</tspan></text>`
-        )
-        currentY += lineHeight
-
-        const mainItems: Array<{ label: string; val: string }> = []
-
-        if (showUptime) {
-          const uptimeVal = (cfg.customUptime as string) || formatUptime(data.user.created_at)
-          if (uptimeVal) mainItems.push({ label: '. Uptime: ', val: ` ${uptimeVal}` })
-        }
-        if (showLocation) {
-          const locVal = (cfg.customLocation as string) || data.user.location
-          if (locVal) mainItems.push({ label: '. Location: ', val: ` ${locVal}` })
-        }
-        if (showCompany) {
-          const compVal = (cfg.customCompany as string) || data.user.company
-          if (compVal) mainItems.push({ label: '. Company: ', val: ` ${compVal}` })
-        }
-        if (showLanguages) {
-          const topLangs =
-            (cfg.customLanguages as string) ||
-            (Object.keys(data.languages).length > 0
-              ? Object.keys(data.languages).slice(0, 5).join(', ')
-              : '')
-          if (topLangs) mainItems.push({ label: '. Languages: ', val: ` ${topLangs}` })
-        }
-        if (showJoined) {
-          const joinedVal =
-            (cfg.customJoined as string) ||
-            (data.user.created_at
-              ? new Date(data.user.created_at).toLocaleDateString('en-US', {
-                  month: 'short',
-                  year: 'numeric',
-                })
-              : '')
-          if (joinedVal) mainItems.push({ label: '. Joined: ', val: ` ${joinedVal}` })
-        }
-        if (showStatus) {
-          const statusVal = cfg.customStatus as string
-          if (statusVal) mainItems.push({ label: '. Status: ', val: ` ${statusVal}` })
-        }
-        if (showPronouns) {
-          const pronounsVal = cfg.customPronouns as string
-          if (pronounsVal) mainItems.push({ label: '. Pronouns: ', val: ` ${pronounsVal}` })
-        }
-        if (showTimezone) {
-          const tzVal = cfg.customTimezone as string
-          if (tzVal) mainItems.push({ label: '. Timezone: ', val: ` ${tzVal}` })
-        }
-        if (showAchievements) {
-          const achVal = cfg.customAchievements as string
-          if (achVal) mainItems.push({ label: '. Achievements: ', val: ` ${achVal}` })
-        }
-        if (showHighlights) {
-          const hlVal = cfg.customHighlights as string
-          if (hlVal) mainItems.push({ label: '. Highlights: ', val: ` ${hlVal}` })
-        }
-
-        for (const item of mainItems) {
-          const dotCount = dotLeaders
-            ? Math.max(2, totalChars - item.label.length - item.val.length)
-            : 2
-          const dotsStr = dotLeaders ? '.'.repeat(dotCount) : '  '
-
-          lines.push(
-            `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${labelClr}">${escapeXml(item.label)}</tspan><tspan fill="${dotClr}">${dotsStr}</tspan><tspan fill="${valClr}">${escapeXml(item.val)}</tspan></text>`
-          )
-          currentY += lineHeight
-        }
-        currentY += 12
-      }
-
-      if (showContactSection) {
-        const contactTitleStr = (cfg.customContactTitle as string) || 'Contact'
-        const dashesCount = Math.max(2, totalChars - 1 - (contactTitleStr.length + 2))
-        const dashesStr = '─'.repeat(dashesCount)
-
-        lines.push(
-          `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${dividerClr}">─</tspan><tspan fill="${headerClr}"> ${escapeXml(contactTitleStr)} </tspan><tspan fill="${dividerClr}">${dashesStr}</tspan></text>`
-        )
-        currentY += lineHeight
-
-        const contactItems: Array<{ label: string; val: string }> = []
-
-        if (showWebsite) {
-          const webVal = (cfg.customWebsite as string) || data.user.blog
-          if (webVal) contactItems.push({ label: '. Website: ', val: ` ${webVal}` })
-        }
-        if (showGithub) {
-          const ghVal = (cfg.customGithub as string) || `github.com/${data.user.login}`
-          if (ghVal) contactItems.push({ label: '. GitHub: ', val: ` ${ghVal}` })
-        }
-        if (showTwitter) {
-          const twVal =
-            (cfg.customTwitter as string) ||
-            (data.user.twitter_username ? `@${data.user.twitter_username}` : '')
-          if (twVal) contactItems.push({ label: '. Twitter: ', val: ` ${twVal}` })
-        }
-        if (showEmail) {
-          const emVal = (cfg.customEmail as string) || data.user.email
-          if (emVal) contactItems.push({ label: '. Email: ', val: ` ${emVal}` })
-        }
-        if (showOrgs) {
-          const orgsVal = cfg.customOrgs as string
-          if (orgsVal) contactItems.push({ label: '. Orgs: ', val: ` ${orgsVal}` })
-        }
-
-        for (const item of contactItems) {
-          const dotCount = dotLeaders
-            ? Math.max(2, totalChars - item.label.length - item.val.length)
-            : 2
-          const dotsStr = dotLeaders ? '.'.repeat(dotCount) : '  '
-
-          lines.push(
-            `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${labelClr}">${escapeXml(item.label)}</tspan><tspan fill="${dotClr}">${dotsStr}</tspan><tspan fill="${valClr}">${escapeXml(item.val)}</tspan></text>`
-          )
-          currentY += lineHeight
-        }
-        currentY += 12
-      }
-
-      if (showStatsSection) {
-        const statsTitleStr = (cfg.customStatsTitle as string) || 'GitHub Stats'
-        const dashesCount = Math.max(2, totalChars - 1 - (statsTitleStr.length + 2))
-        const dashesStr = '─'.repeat(dashesCount)
-
-        lines.push(
-          `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${dividerClr}">─</tspan><tspan fill="${headerClr}"> ${escapeXml(statsTitleStr)} </tspan><tspan fill="${dividerClr}">${dashesStr}</tspan></text>`
-        )
-        currentY += lineHeight
-
-        const statFields: Array<{ label: string; val: string }> = []
-
-        if (showRepos) {
-          statFields.push({ label: '. Repos: ', val: ` ${data.user.public_repos}` })
-        }
-        if (showStars) {
-          statFields.push({ label: '. Stars: ', val: ` ${data.totalStars}` })
-        }
-        if (showCommits) {
-          const commitCount =
-            (cfg.customCommits as string) ||
-            (data.contributions ? data.contributions.totalContributions.toLocaleString() : '')
-          if (commitCount) statFields.push({ label: '. Commits: ', val: ` ${commitCount}` })
-        }
-        if (showFollowers) {
-          statFields.push({ label: '. Followers: ', val: ` ${data.user.followers}` })
-        }
-        if (showFollowing) {
-          const followingCount =
-            (cfg.customFollowing as string) ||
-            (data.user.following !== undefined ? data.user.following.toLocaleString() : '')
-          if (followingCount) statFields.push({ label: '. Following: ', val: ` ${followingCount}` })
-        }
-        if (showGists) {
-          const gistsCount =
-            (cfg.customGists as string) ||
-            (data.user.public_gists !== undefined ? data.user.public_gists.toLocaleString() : '')
-          if (gistsCount) statFields.push({ label: '. Gists: ', val: ` ${gistsCount}` })
-        }
-
-        for (let i = 0; i < statFields.length; i += 2) {
-          const item1 = statFields[i]
-          const item2 = statFields[i + 1]
-
-          if (item1 && item2) {
-            const halfChars = Math.floor((totalChars - 3) / 2)
-            const dots1 = dotLeaders
-              ? Math.max(2, halfChars - item1.label.length - item1.val.length)
-              : 2
-            const dots2 = dotLeaders
-              ? Math.max(2, halfChars - item2.label.length - item2.val.length)
-              : 2
-
-            lines.push(
-              `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${labelClr}">${escapeXml(item1.label)}</tspan><tspan fill="${dotClr}">${'.'.repeat(dots1)}</tspan><tspan fill="${statsValClr}">${escapeXml(item1.val)}</tspan><tspan fill="${dividerClr}"> | </tspan><tspan fill="${labelClr}">${escapeXml(item2.label)}</tspan><tspan fill="${dotClr}">${'.'.repeat(dots2)}</tspan><tspan fill="${statsValClr}">${escapeXml(item2.val)}</tspan></text>`
-            )
-          } else if (item1) {
-            const dots1 = dotLeaders
-              ? Math.max(2, totalChars - item1.label.length - item1.val.length)
-              : 2
-            lines.push(
-              `<text x="${paddingX}" y="${currentY}" font-family="'Consolas', 'Menlo', 'DejaVu Sans Mono', 'JetBrains Mono', monospace" xml:space="preserve" font-size="${fontSize}"><tspan fill="${labelClr}">${escapeXml(item1.label)}</tspan><tspan fill="${dotClr}">${'.'.repeat(dots1)}</tspan><tspan fill="${statsValClr}">${escapeXml(item1.val)}</tspan></text>`
-            )
-          }
-          currentY += lineHeight
-        }
-      }
-
-      contentSvg = lines.join('\n')
-      break
-    }
-
-    case 'gitfest-lineup': {
-      const username = (cfg.username as string) || data.user.login
-      const theme = (cfg.theme as string) || 'dark'
-      const sort = (cfg.sort as string) || 'stars'
-      const order = (cfg.order as string) || 'asc'
-      const type = (cfg.type as string) || 'owner'
-      const invertColors = Boolean(cfg.invertColors)
-      const hideRepos = (cfg.hideRepos as string) || ''
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ GITFEST LINEUP ]'
-
-      const baseUrl = process.env.NEXT_PUBLIC_GITFEST_URL || 'http://localhost:3000'
-      const statsUrl = `${baseUrl}/api/lineup?username=${encodeURIComponent(username)}&theme=${theme}&sort=${sort}&order=${order}&type=${type}${invertColors ? '&invertColors=true' : ''}${hideRepos ? `&hideRepos=${encodeURIComponent(hideRepos)}` : ''}`
-
-      contentSvg = renderExternalWidgetSvg(
-        statsUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        globalStyles.accentColor || '',
-        'contain'
-      )
-      break
-    }
-
-    case 'github-readme-stats': {
-      const username = (cfg.username as string) || data.user.login
-      const statType = (cfg.statType as string) || 'stats'
-      const theme = (cfg.theme as string) || 'dark'
-      const showIcons = cfg.showIcons !== false
-      const countPrivate = Boolean(cfg.countPrivate)
-      const includeAllCommits = Boolean(cfg.includeAllCommits)
-      const hideRank = Boolean(cfg.hideRank)
-      const hideBorder = Boolean(cfg.hideBorder)
-      const repoName = (cfg.repoName as string) || data.repos[0]?.name || 'gitascii'
-      const layout = (cfg.layout as string) || 'compact'
-      const langsCount = Number(cfg.langsCount) || 5
-      const hideLangs = (cfg.hideLangs as string) || ''
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ GITHUB README STATS ]'
-
-      let statsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.README_STATS(
-        username,
-        `&show_icons=${showIcons}&theme=${theme}${countPrivate ? '&count_private=true' : ''}${includeAllCommits ? '&include_all_commits=true' : ''}${hideRank ? '&hide_rank=true' : ''}${hideBorder ? '&hide_border=true' : ''}`
-      )
-      if (statType === 'top-langs') {
-        statsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.TOP_LANGS(
-          username,
-          `&layout=${layout}&langs_count=${langsCount}&theme=${theme}${hideLangs ? `&hide=${encodeURIComponent(hideLangs)}` : ''}${hideBorder ? '&hide_border=true' : ''}`
-        )
-      } else if (statType === 'pin') {
-        statsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.PIN_REPO(
-          username,
-          repoName,
-          `&theme=${theme}${hideBorder ? '&hide_border=true' : ''}`
-        )
-      }
-
-      contentSvg = renderExternalWidgetSvg(
-        statsUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'godprofile-terminal': {
-      contentSvg = renderTerminal(widget, data, globalStyles)
-      break
-    }
-
-    case 'godprofile-marquee': {
-      contentSvg = renderMarquee(widget, data, globalStyles)
-      break
-    }
-
-    case 'godprofile-neural': {
-      contentSvg = renderNeural(widget, data, globalStyles)
-      break
-    }
-
-    case 'godprofile-trophies': {
-      contentSvg = renderTrophies(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.GODPROFILE_WAKATIME: {
-      contentSvg = renderWakaTime(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.GODPROFILE_GLOBE: {
-      contentSvg = renderGlobe(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.ASCII_PORTRAIT: {
-      contentSvg = renderAsciiPortrait(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.ASCII_INFO: {
-      contentSvg = renderAsciiInfoCard(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.ASCII_HEATMAP: {
-      contentSvg = renderAsciiHeatmap(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_SYSTEM_LOOP: {
-      contentSvg = renderSystemLoop(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_COMMAND_DECK: {
-      contentSvg = renderCommandDeck(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_SIGNAL_GRID: {
-      contentSvg = renderSignalGrid(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_METRO: {
-      contentSvg = renderMetroMap(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_BENTO: {
-      contentSvg = renderBentoGrid(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_EDITORIAL: {
-      contentSvg = renderEditorial(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_BLUEPRINT: {
-      contentSvg = renderBlueprint(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_CONSTELLATION: {
-      contentSvg = renderConstellation(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_MONOLITH: {
-      contentSvg = renderMonolith(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_INTERLACE: {
-      contentSvg = renderInterlace(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_CIPHER: {
-      contentSvg = renderCipherPrint(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_SPECIMEN: {
-      contentSvg = renderFieldSpecimen(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_PATCHBAY: {
-      contentSvg = renderPatchbay(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_CARTOGRAPH: {
-      contentSvg = renderCartograph(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CONTROLPLANE_FOUNDRY: {
-      contentSvg = renderFoundry(widget, data, globalStyles)
-      break
-    }
-
-    case WIDGET_IDS.CODEWEB_HERO_ORBIT: {
-      contentSvg = renderCodewebHeroOrbit(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.CODEWEB_RETRO_GRID: {
-      contentSvg = renderCodewebRetroGrid(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.CODEWEB_SHOWCASE_CARDS: {
-      contentSvg = renderCodewebShowcaseCards(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.CODEWEB_SOCIAL_BADGE: {
-      contentSvg = renderCodewebSocialBadge(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case WIDGET_IDS.CODEWEB_MINIMAL_BADGE: {
-      contentSvg = renderCodewebMinimalBadge(widget, data, globalStyles, forceStatic)
-      break
-    }
-
-    case 'ghstats': {
-      const username = data.user.login
-      const embedType = (cfg.embedType as string) || 'card'
-
-      const theme = (cfg.theme as string) || 'default'
-      const showIcons = cfg.showIcons !== false
-      const showRing = cfg.showRing !== false
-      const hideBorder = Boolean(cfg.hideBorder)
-      const hideTitle = Boolean(cfg.hideTitle)
-      const size = (cfg.size as string) || 'default'
-      const compactCount = (cfg.compactCount as string) || '4'
-      const hideStats = (cfg.hideStats as string) || ''
-
-      const customTitle = (cfg.customTitle as string) || ''
-      const layout = (cfg.layout as string) || 'bar'
-      const maxLangs = Number(cfg.maxLangs) || 8
-      const badgeStyle = (cfg.badgeStyle as string) || 'flat'
-
-      let statsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.GH_STATS(embedType, username, theme)
-
-      const bgColor = cfg.backgroundColor as string
-      if (bgColor) statsUrl += `&bg=${bgColor.replace('#', '')}`
-
-      const textColor = cfg.textColor as string
-      if (textColor) statsUrl += `&text=${textColor.replace('#', '')}`
-
-      const accentColor = cfg.accentColor as string
-      if (accentColor) {
-        statsUrl += `&icon_color=${accentColor.replace('#', '')}`
-        statsUrl += `&title_color=${accentColor.replace('#', '')}`
-      }
-
-      const borderColor = cfg.borderColor as string
-      if (borderColor) statsUrl += `&border_color=${borderColor.replace('#', '')}`
-
-      if (embedType === 'card') {
-        if (!showIcons) statsUrl += `&show_icons=false`
-        if (!showRing) statsUrl += `&show_ring=false`
-        if (hideBorder) statsUrl += `&hide_border=true`
-        if (hideTitle) statsUrl += `&hide_title=true`
-        if (size === 'compact') statsUrl += `&size=compact&compact_count=${compactCount}`
-        if (customTitle) statsUrl += `&custom_title=${encodeURIComponent(customTitle)}`
-        if (hideStats) statsUrl += `&hide=${encodeURIComponent(hideStats)}`
-      } else if (embedType === 'langs') {
-        if (layout !== 'bar') statsUrl += `&layout=${layout}`
-        if (maxLangs !== 8) statsUrl += `&max_langs=${maxLangs}`
-        if (hideBorder) statsUrl += `&hide_border=true`
-      } else if (embedType === 'mini' || embedType === 'badge') {
-        if (badgeStyle !== 'flat') statsUrl += `&style=${badgeStyle}`
-      }
-
-      contentSvg = renderExternalWidgetSvg(
-        statsUrl,
-        width,
-        height,
-        customTitle || '[ GHSTATS.DEV ]',
-        !hideTitle && embedType !== 'card',
-        globalStyles,
-        accent,
-        'contain'
-      )
-
-      break
-    }
-
-    case 'streak-stats': {
-      const username = (cfg.username as string) || data.user.login
-      const theme = (cfg.theme as string) || 'dark'
-      const mode = (cfg.mode as string) || 'daily'
-      const dateFormat = (cfg.dateFormat as string) || 'M j, Y'
-      const streakBorderRadius = Number(cfg.streakBorderRadius) || 4
-      const hideBorder = Boolean(cfg.hideBorder)
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ GITHUB STREAK STATS ]'
-
-      const streakUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.STREAK_STATS(
-        username,
-        `&theme=${theme}&mode=${mode}&date_format=${encodeURIComponent(dateFormat)}&border_radius=${streakBorderRadius}${hideBorder ? '&hide_border=true' : ''}`
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        streakUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'profile-trophy': {
-      const username = (cfg.username as string) || data.user.login
-      const theme = (cfg.theme as string) || 'flat'
-      const column = Number(cfg.column) || 6
-      const row = Number(cfg.row) || 1
-      const noFrame = Boolean(cfg.noFrame)
-      const noBg = Boolean(cfg.noBg)
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ PROFILE TROPHIES ]'
-
-      const trophyUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.PROFILE_TROPHY(
-        username,
-        `&theme=${theme}&column=${column}&row=${row}${noFrame ? '&margin-w=0' : ''}${noBg ? '&no-bg=true' : ''}`
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        trophyUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'activity-graph': {
-      const username = (cfg.username as string) || data.user.login
-      const theme = (cfg.theme as string) || 'github-dark'
-      const days = Number(cfg.days) || 31
-      const showArea = cfg.showArea !== false
-      const hideBorder = Boolean(cfg.hideBorder)
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ ACTIVITY GRAPH ]'
-
-      const graphUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.ACTIVITY_GRAPH(
-        username,
-        `&theme=${theme}&days=${days}&area=${showArea}${hideBorder ? '&hide_border=true' : ''}`
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        graphUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'contribution-snake': {
-      const username = (cfg.username as string) || data.user.login
-      const theme = (cfg.theme as string) || 'dark'
-      const branch = (cfg.branch as string) || 'output'
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ CONTRIBUTION SNAKE ]'
-
-      const snakeFileName =
-        theme === 'light'
-          ? 'github-contribution-grid-snake.svg'
-          : 'github-contribution-grid-snake-dark.svg'
-      const snakeUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.JSDELIVR_GH(
-        username,
-        username,
-        branch,
-        snakeFileName
-      )
-      const fallbackSnakeUrl =
-        API_ENDPOINTS.EXTERNAL_RESOURCES.PLATANE_FALLBACK_SNAKE(snakeFileName)
-
-      contentSvg = renderExternalWidgetSvg(
-        snakeUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain',
-        undefined,
-        fallbackSnakeUrl
-      )
-      break
-    }
-
-    case 'metrics-card': {
-      const username = (cfg.username as string) || data.user.login
-      const template = (cfg.template as string) || 'classic'
-      const baseSections = (cfg.baseSections as string) || 'header,activity,community,repositories'
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ GITHUB METRICS CARD ]'
-
-      const metricsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.LECOQ_METRICS(
-        username,
-        template,
-        baseSections
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        metricsUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'views-counter': {
-      const username = (cfg.username as string) || data.user.login
-      const color = (cfg.color as string) || '00f0ff'
-      const style = (cfg.style as string) || 'for-the-badge'
-      const label = (cfg.label as string) || 'PROFILE VIEWS'
-      const baseVal = Number(cfg.baseVal) || 0
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ VIEWS COUNTER ]'
-
-      const viewsUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.KOMAREV_VIEWS(
-        username,
-        `&color=${color}&style=${style}&label=${encodeURIComponent(label)}${baseVal > 0 ? `&base=${baseVal}` : ''}`
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        viewsUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'badge'
-      )
-      break
-    }
-
-    case 'readme-quotes': {
-      const quoteType = (cfg.quoteType as string) || 'random'
-      const theme = (cfg.theme as string) || 'dark'
-      const layout = (cfg.layout as string) || 'horizontal'
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ DEVELOPER QUOTE ]'
-
-      const quoteUrl = API_ENDPOINTS.EXTERNAL_WIDGETS.QUOTES(
-        quoteType === 'quote-day' ? 'quote-day' : layout,
-        theme
-      )
-
-      contentSvg = renderExternalWidgetSvg(
-        quoteUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'contain'
-      )
-      break
-    }
-
-    case 'awesome-badge': {
-      const badgeStyle = (cfg.badgeStyle as string) || 'for-the-badge'
-      const badgeColor = (cfg.badgeColor as string) || 'brightgreen'
-      const label = (cfg.label as string) || 'Awesome GitHub Profile'
-      const logo = (cfg.logo as string) || 'github'
-
-      const showTitle = cfg.showTitle !== false
-      const customTitle = (cfg.customTitle as string) || '[ AWESOME PROFILE BADGE ]'
-
-      const badgeUrl = API_ENDPOINTS.SHIELDS_IO.BADGE(
-        label,
-        'Featured',
-        badgeColor,
-        badgeStyle,
-        logo
-      )
-      const targetUrl = EXTERNAL_LINKS.AWESOME_GITHUB_PROFILE_README
-
-      contentSvg = renderExternalWidgetSvg(
-        badgeUrl,
-        width,
-        height,
-        customTitle,
-        showTitle,
-        globalStyles,
-        accent,
-        'badge',
-        targetUrl
-      )
-      break
-    }
-
-    case 'custom-image': {
-      const imageUrl = (cfg.imageUrl as string) || (cfg.src as string) || (cfg.url as string) || ''
-      const targetUrl = (cfg.targetUrl as string) || (cfg.href as string) || undefined
-      const showTitle = cfg.showTitle === true
-      const customTitle = (cfg.customTitle as string) || '[ IMAGE ]'
-      const mode = (cfg.mode as 'contain' | 'badge') || 'contain'
-
-      if (!imageUrl) {
-        contentSvg = `
-          <rect width="${width}" height="${height}" fill="#18181b" rx="4" opacity="0.6" stroke="${border}" stroke-width="1" />
-          <text x="${width / 2}" y="${height / 2 - 6}" text-anchor="middle" font-family="${globalStyles.fontFamily}" font-size="12" fill="${accent}">📷 [ IMAGEM CUSTOMIZADA ]</text>
-          <text x="${width / 2}" y="${height / 2 + 14}" text-anchor="middle" font-family="${globalStyles.fontFamily}" font-size="10" fill="#71717a">Cole a URL ou faça upload no painel de propriedades</text>
-        `
-      } else {
-        contentSvg = renderExternalWidgetSvg(
-          imageUrl,
-          width,
-          height,
-          customTitle,
-          showTitle,
-          globalStyles,
-          accent,
-          mode,
-          targetUrl
-        )
-      }
-      break
-    }
-
-    default: {
-      contentSvg = `
-        <text x="24" y="36" font-family="'Inter Tight', sans-serif" font-size="14" fill="${textClr}">${escapeXml(widget.widgetId.toUpperCase())}</text>
-      `
-    }
-  }
+  const renderer = getRenderer(widget.widgetId)
+  let contentSvg = renderer(widget, data, globalStyles, forceStatic)
 
   let templateDecorationSvg = ''
   const tmplStyle = globalStyles.templateStyle || 'terminal'
