+++ src/engine/core/xmlUtils.ts
@@ -1,0 +1,42 @@
+/**
+ * XML and SVG formatting utilities for GitAscii renderers.
+ */
+
+export function escapeXml(str: string): string {
+  if (!str) return ''
+  return str
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&apos;')
+}
+
+export function formatUptime(createdAtStr?: string): string {
+  if (!createdAtStr) return '5 years, 3 months, 13 days'
+  const created = new Date(createdAtStr)
+  if (isNaN(created.getTime())) return '5 years, 3 months, 13 days'
+  const now = new Date()
+
+  let years = now.getFullYear() - created.getFullYear()
+  let months = now.getMonth() - created.getMonth()
+  let days = now.getDate() - created.getDate()
+
+  if (days < 0) {
+    months -= 1
+    const prevMonth = new Date(now.getFullYear(), now.getMonth(), 0)
+    days += prevMonth.getDate()
+  }
+
+  if (months < 0) {
+    years -= 1
+    months += 12
+  }
+
+  const parts = []
+  if (years > 0) parts.push(`${years} ${years === 1 ? 'year' : 'years'}`)
+  if (months > 0) parts.push(`${months} ${months === 1 ? 'month' : 'months'}`)
+  if (days > 0 || parts.length === 0) parts.push(`${days} ${days === 1 ? 'day' : 'days'}`)
+
+  return parts.join(', ')
+}
