+++ src/features/editor/components/Canvas/CanvasStatusBar.tsx
@@ -13,12 +13,17 @@ interface CanvasStatusBarProps {
 
 export function CanvasStatusBar({ showInfo = true }: CanvasStatusBarProps) {
   const { t } = useI18n()
-  const { zoom, setZoom, config, selectedInstanceIds, undo, redo, canUndo, canRedo } =
-    useEditorStore()
+
+  const zoom = useEditorStore((state) => state.zoom)
+  const setZoom = useEditorStore((state) => state.setZoom)
+  const widgetCount = useEditorStore((state) => state.config?.widgets?.length ?? 0)
+  const selCount = useEditorStore((state) => state.selectedInstanceIds.length)
+  const undo = useEditorStore((state) => state.undo)
+  const redo = useEditorStore((state) => state.redo)
+  const canUndo = useEditorStore((state) => state.canUndo)
+  const canRedo = useEditorStore((state) => state.canRedo)
 
   const zoomPct = Math.round(zoom * 100)
-  const widgetCount = config?.widgets?.length ?? 0
-  const selCount = selectedInstanceIds?.length ?? 0
 
   const stepDown = () => setZoom(Math.max(0.5, zoom - 0.1))
   const stepUp = () => setZoom(Math.min(1.5, zoom + 0.1))
@@ -63,7 +68,7 @@ export function CanvasStatusBar({ showInfo = true }: CanvasStatusBarProps) {
         )}
       </div>
 
-      <div className="flex items-center gap-0.5">
+      <div className="flex items-center gap-1">
         <button
           onClick={stepDown}
           title={t('editor.statusbar.zoom_out')}
@@ -73,14 +78,12 @@ export function CanvasStatusBar({ showInfo = true }: CanvasStatusBarProps) {
           <Minus size={13} />
         </button>
 
-        <button
-          onClick={resetZoom}
-          title={t('editor.statusbar.zoom_reset')}
-          data-testid="statusbar-zoom-reset"
-          className="min-w-11.5 text-center text-pearl hover:text-chalk hover:bg-graphite rounded-xs px-1.5 py-0.5 transition-colors cursor-pointer font-jetbrains-mono text-[11px] font-semibold"
+        <span
+          data-testid="statusbar-zoom-level"
+          className="w-11 text-center font-mono font-medium text-chalk text-caption"
         >
           {zoomPct}%
-        </button>
+        </span>
 
         <button
           onClick={stepUp}
@@ -91,13 +94,11 @@ export function CanvasStatusBar({ showInfo = true }: CanvasStatusBarProps) {
           <Plus size={13} />
         </button>
 
-        <div className="w-px h-3 bg-graphite mx-1" />
-
         <button
           onClick={resetZoom}
           title={t('editor.statusbar.zoom_reset')}
-          data-testid="statusbar-zoom-fit"
-          className="p-1 rounded-xs text-ash hover:bg-graphite hover:text-chalk transition-colors cursor-pointer"
+          data-testid="statusbar-zoom-reset"
+          className="p-1 rounded-xs text-ash hover:bg-graphite hover:text-chalk transition-colors cursor-pointer ml-0.5"
         >
           <RotateCcw size={12} />
         </button>
