+++ src/features/widgets/renderers/DividerFooterRenderer.ts
@@ -1,0 +1,30 @@
+import { APP_DOMAIN } from '@/constants'
+import type { GlobalStyles, NormalizedGitHubData, WidgetInstance } from '@/engine/types'
+
+export function renderDivider(
+  widget: WidgetInstance,
+  _data: NormalizedGitHubData,
+  globalStyles: GlobalStyles
+): string {
+  const { width, height } = widget.size
+  const cfg = widget.config
+  const accent = (cfg.accentColor as string) || globalStyles.accentColor || '#c5ff4a'
+
+  return `<line x1="0" y1="${height / 2}" x2="${width}" y2="${height / 2}" stroke="${accent}" stroke-width="4" />`
+}
+
+export function renderFooter(
+  widget: WidgetInstance,
+  data: NormalizedGitHubData,
+  globalStyles: GlobalStyles
+): string {
+  const { width, height } = widget.size
+  const cfg = widget.config
+  const accent = (cfg.accentColor as string) || globalStyles.accentColor || '#c5ff4a'
+
+  return `
+    <rect x="0" y="0" width="${width}" height="${height}" fill="#000000" />
+    <text x="24" y="${height / 2 + 4}" font-family="'Inter Tight', sans-serif" font-size="11" font-weight="500" fill="#7a7a7a" letter-spacing="2">[ GENERATED BY GITASCII ]</text>
+    <text x="${width - 24}" y="${height / 2 + 4}" text-anchor="end" font-family="'JetBrains Mono', monospace" font-size="11" fill="${accent}">${APP_DOMAIN}/${data.user.login}</text>
+  `
+}
